# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "LINUX DO Credit",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/order/transactions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/order.TransactionListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "order.TransactionListRequest": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer",
                    "minimum": 1
                },
                "page_size": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "payee_username": {
                    "type": "string"
                },
                "payer_username": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "success",
                        "pending",
                        "failed",
                        "expired",
                        "disputing",
                        "refund",
                        "refused"
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "receive",
                        "payment",
                        "transfer",
                        "community",
                        "online",
                        "test"
                    ]
                }
            }
        },
...
        "util.ResponseAny": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    }
}